% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Celsius_to_Fahrenheit.R
\name{Celsius_to_Fahrenheit}
\alias{Celsius_to_Fahrenheit}
\title{Converts Temperature in Celsius to Fahrenheit}
\usage{
Celsius_to_Fahrenheit(temp_C)
}
\arguments{
\item{temp_C}{numeric}
}
\value{
numeric
}
\description{
Converts Temperature in Celsius to Fahrenheit
}
\examples{
Celsius_to_Fahrenheit(1) # results in 33.8F
}
