% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Celsius_to_Kelvin.R
\name{Celsius_to_Kelvin}
\alias{Celsius_to_Kelvin}
\title{Converts Temperature from Celsius to Kelvin}
\usage{
Celsius_to_Kelvin(temp_C)
}
\arguments{
\item{temp_C}{numeric}
}
\value{
numeric
}
\description{
Converts Temperature from Celsius to Kelvin
}
\examples{
Celsius_to_Kelvin(1) # returns 274.15K
}
