% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fahrenheit_to_Celsius.R
\name{Fahrenheit_to_Celsius}
\alias{Fahrenheit_to_Celsius}
\title{Converts Temperature from Fahrenheit to Celsius}
\usage{
Fahrenheit_to_Celsius(temp_F)
}
\arguments{
\item{temp_F}{numeric}
}
\value{
numeric
}
\description{
Converts Temperature from Fahrenheit to Celsius
}
\examples{
Fahrenheit_to_Celsius(1) #returns -17.22222
}
