% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fahrenheit_to_Kelvin.R
\name{Fahrenheit_to_Kelvin}
\alias{Fahrenheit_to_Kelvin}
\title{Converts Temperature from Fahrenheit to Kelvin}
\usage{
Fahrenheit_to_Kelvin(temp_F)
}
\arguments{
\item{temp_F}{numeric}
}
\value{
numeric
}
\description{
Converts Temperature from Fahrenheit to Kelvin
}
\examples{
Fahrenheit_to_Kelvin(1) #returns 255.9278K
}
