% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kelvin_to_Celsius.R
\name{Kelvin_to_Celsius}
\alias{Kelvin_to_Celsius}
\title{Converts Temperature in Kelving to Celsius}
\usage{
Kelvin_to_Celsius(temp_K)
}
\arguments{
\item{temp_K}{numeric}
}
\value{
numeric
}
\description{
Converts Temperature in Kelving to Celsius
}
\examples{
Kelvin_to_Celsius(1) # returns -272.15
}
