% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kelvin_to_Fahrenheit.R
\name{Kelvin_to_Fahrenheit}
\alias{Kelvin_to_Fahrenheit}
\title{Converts Temperature in Kelvin to Fahrenheit}
\usage{
Kelvin_to_Fahrenheit(temp_K)
}
\arguments{
\item{temp_K}{numeric}
}
\value{
numeric
}
\description{
Converts Temperature in Kelvin to Fahrenheit
}
\examples{
Kelvin_to_Fahrenheit(255.9278) # returns 1F
}
